+++ what4/src/What4/Protocol/Online.hs
@@ -32,12 +32,15 @@ module What4.Protocol.Online
   , reset
   , inNewFrame
   , inNewFrameWithVars
+  , inNewFrame2Open
+  , inNewFrame2Close
   , check
   , checkAndGetModel
   , checkWithAssumptions
   , checkWithAssumptionsAndModel
   , getModel
   , getUnsatCore
+  , getAbducts
   , getUnsatAssumptions
   , getSatResult
   , checkSatisfiable
@@ -236,6 +239,39 @@ checkSatisfiable proc rsn p =
         do assume conn p
            check proc rsn
 
+-- | @get-abuct nm t@ queries the solver for the first abduct, which is returned
+--   as an SMT function definition named @nm@. The remaining abducts are obtained
+--   from the solver by successive invocations of the @get-abduct-next@ command,
+--   which return SMT functions bound to the same @nm@ as the first. The name @nm@
+--   is bound within the current assertion frame.
+--   Note that this is an unstable API; we expect that the return type will change 
+--   to a parsed expression in the future
+getAbducts ::
+  SMTReadWriter solver =>
+  SolverProcess scope solver ->
+  Int ->
+  Text ->
+  BoolExpr scope ->
+  IO [String]
+getAbducts proc n nm t =
+  if (n > 0) then do 
+    let conn = solverConn proc
+    unless (supportedFeatures conn `hasProblemFeature` useProduceAbducts) $
+      fail $ show $ pretty (smtWriterName conn) <+> pretty "is not configured to produce abducts"
+    f <- mkFormula conn t
+    -- get the first abduct using the get-abduct command
+    addCommandNoAck conn (getAbductCommand conn nm f)
+    abd1 <- smtAbductResult conn conn nm f
+    -- get the remaining abducts using get-abduct-next commands
+    if (n > 1) then do
+      let rest = n - 1
+      abdRest <- forM [1..rest] $ \_ -> do
+        addCommandNoAck conn (getAbductNextCommand conn)
+        smtAbductNextResult conn conn
+      return (abd1:abdRest)
+    else return [abd1]
+  else return []
+
 -- | Check if the formula is satisifiable in the current
 --   solver state.  This is done in a
 --   fresh frame, which is exited after the continuation
@@ -326,6 +362,17 @@ tryPop p =
 -- | Perform an action in the scope of a solver assumption frame.
 inNewFrame :: (MonadIO m, MonadMask m, SMTReadWriter solver) => SolverProcess scope solver -> m a -> m a
 inNewFrame p action = inNewFrameWithVars p [] action
+
+-- | Open a second solver assumption frame.
+-- For abduction, we want the final assertion to be a in a new frame, so that it 
+-- can be closed before asking for abducts. The following two commands allow frame 2 
+-- to be pushed and popped independently of other commands
+inNewFrame2Open :: SMTReadWriter solver => SolverProcess scope solver -> IO ()
+inNewFrame2Open sp = let c = solverConn sp in addCommand c (push2Command c)
+
+-- | Close a second solver assumption frame.
+inNewFrame2Close :: SMTReadWriter solver => SolverProcess scope solver -> IO ()
+inNewFrame2Close sp = let c = solverConn sp in addCommand c (pop2Command c)
 
 -- | Perform an action in the scope of a solver assumption frame, where the given
 -- bound variables are considered free within that frame.