+++ what4/src/What4/Solver/Yices.hs
@@ -331,6 +331,9 @@ instance SupportTermOps YicesTerm where
 
   fromText t = T (Builder.fromText t)
 
+unsupportedFeature :: String -> a
+unsupportedFeature s = error ("Yices does not support " <> s)
+
 floatFail :: HasCallStack => a
 floatFail = error "Yices does not support IEEE-754 floating-point numbers"
 
@@ -492,6 +495,8 @@ instance SMTWriter Connection where
 
   pushCommand _   = const $ safeCmd "(push)"
   popCommand _    = const $ safeCmd "(pop)"
+  push2Command _   = unsupportedFeature "(push 2)"
+  pop2Command _    = unsupportedFeature "(pop 2)"
   resetCommand _  = const $ safeCmd "(reset)"
   checkCommands _  =
     [ setTimeoutCommand, const $ safeCmd "(check)" ]
@@ -502,6 +507,9 @@ instance SMTWriter Connection where
 
   getUnsatAssumptionsCommand _ = const $ safeCmd "(show-unsat-assumptions)"
   getUnsatCoreCommand _ = const $ safeCmd "(show-unsat-core)"
+  getAbductCommand _ _ _ = unsupportedFeature "abduction"
+  getAbductNextCommand _ = unsupportedFeature "abduction"
+
   setOptCommand _ x o = setParamCommand x (Builder.fromText o)
 
   assertCommand _ (T nm) = const $ unsafeCmd $ app "assert" [nm]
@@ -589,6 +597,9 @@ instance SMTReadWriter Connection where
                  unlines [ "Could not parse unsat core result."
                          , "*** Exception: " ++ displayException e
                          ]
+  smtAbductResult _ _ _ = unsupportedFeature "abduction"
+
+  smtAbductNextResult _ = unsupportedFeature "abduction"
 
 
 -- | Exceptions that can occur when reading responses from Yices